### 2023-12-19 #6152 @Kevin-Mattheus-Moerman
@katjaq thanks for this submission. I am the AEiC on this track and here to help with initial steps.

For the moment I am trying to determine if this work is in scope for JOSS. Could you help clarify if the above software report is accurate :point_up:. Would it be correct to say that your creative contribution is captures by the 2476 lines of JavaScript? If so would you say all of that is your work or is some of it content from other libraries or auto-generated?
