### 2023-12-19 #6152 @katjaq
Hi @Kevin-Mattheus-Moerman . Thank you very much for your message.

Our creative contribution includes 2816 lines of Javascript. This combines the code directly included within the Thresholdmann repo, as well as code that we wrote, but include here as a module we call from jsdelivr.

Within our Thresholdmann repo, all code except for one file is exclusively our contribution:
* thresholdmann.js : 829 lines
* worker.js : 88 lines
* test.mjs : 144 lines
* render3D/index.js : 113 lines
* surfacenets.worker.js : about 20 lines from 151 are from us, and some adaptations

The code we wrote and include here in form of modules includes
* mui.js : 250 lines
* mriviewer.js : 831 lines, which are both called from Thresholdmann/index.html, and
* mri.js : 541 lines, called from mriviewer.js.

So, 829+88+144+113+20+250+541+831 = 2816 lines our contribution. plus 284 lines in our [eslint config file](https://github.com/neuroanatomy/eslint-config-naat/blob/main/index.js) which we wrote to enforce our coding style (across all our code). But we didn't count that here as we are not sure how you consider eslint :)

Thresholdmann css and html are also important for the functioning of our GUI. Together with the documentation html, they add up to additional 778 lines of original contribution.

Thank you for your consideration.
