### 2023-12-19 #6152 @katjaq
For potential reviewers, we checked the JOSS list, and thought that Katie Bottenhorn (62442katieb), Anibal Sólon (anibalsolon) and Jonny Saunders (sneakers-the-rat) could be potential great reviewers. :)
