### 2024-01-09 #6152 @r03ert0
hi @Kevin-Mattheus-Moerman,
did you reach a conclusion on whether our work fulfils the criteria for JOSS?
