### 2024-01-15 #6152 @emdupre
Hi @Kevin-Mattheus-Moerman, thanks for the invitation !

Just to be transparent : I've coauthored with both the first and senior author for this submission on [this 2021 paper ](https://pubmed.ncbi.nlm.nih.gov/34414422/), which would qualify under [the COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). Although I had worked closely with these authors in a volunteer committee through 2020, we have had no formal or informal affiliation since.
