### 2024-01-21 #6152 @Kevin-Mattheus-Moerman
@katjaq thanks for your message. Unfortunately you can't help here. I'm aware of the potentially suitable editors but unfortunately they are all busy handling other submissions. We will assign an editor once one becomes available. We are also recruiting new editors at the moment who may be able to help. I hope we'll be able to assign an editor within 2 weeks.
