### 2024-02-06 #6152 @adamltyson
Hi @katjaq, thanks for the reviewer suggestions!

@62442katieb, @anibalsolon, @sneakers-the-rat would you be able to review this submission to JOSS? I think you've all reviewed for JOSS before, but if you have any questions about the process, please let me know. Thanks!
