### 2024-02-07 #6152 @sneakers-the-rat
hello! i always love a reason to take a look at what @katjaq has been up to.

I have also been recruited as an editor but have not yet been able to finish my onboarding because i got pulled between several things since then, so i'm not sure if that's a problem or not. if it's not, then yes! would love to! i'll be free to start next week.
