### 2024-02-07 #6152 @danielskatz
> hello! i always love a reason to take a look at what @katjaq has been up to.
>
> I have also been recruited as an editor but have not yet been able to finish my onboarding because i got pulled between several things since then, so i'm not sure if that's a problem or not. if it's not, then yes! would love to! i'll be free to start next week.
>
> edit: one more thing, i have reviewed a previous JOSS submission from this group ( #2670 ) I don't know if that's a COI (it doesn't feel like one) but i figured i'd ask

In my opinion as an associate-editor-in-chief, an editor should dedicate most of their JOSS time to editing, not reviewing, but if you haven't started editing yet, it seems reasonable for you to be a reviewer.
