### 2024-02-08 #6336 @editorialbot
[REVIEW]: Thresholdmann: A Web tool for interactively creating adaptive thresholds to segment MRI data.
**Submitting author:** @katjaq (<a href="http://orcid.org/0000-0002-7237-0196">Katja Heuer</a>)
**Repository:** https://github.com/neuroanatomy/thresholdmann
**Branch with paper.md** (empty if default branch): joss-paper-branch
**Version:** v.1.0.0
