### 2024-02-08 #6336 @adamltyson
👋🏼 @katjaq @sneakers-the-rat @62442katieb @anibalsolon, this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
