### 2024-02-19 neuroanatomy/thresholdmann#16 @sneakers-the-rat
[joss][nice to have] Provide demo data?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Hello!!

as usual I like the dead simple "drag a nifti onto this big box" interface, and I see the `bear_uchar.nii.gz` used in the tests as example data. For the sake of reviewers and also for potential users, would it be possible to have a "demo mode" that loads that example file from the UI so folks can test it out without needing to have a nifti on hand?

edit: [brainbox](https://brainbox.pasteur.fr/) has a great example of what i mean with the demo data selector :).
