### 2024-02-19 neuroanatomy/thresholdmann#17 @sneakers-the-rat
[joss][docs] Developer docs
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Collecting ideas for developer docs here rather than in separate issues. Trying to minimize nags and only focus on things that I think would be helpful for inviting contributors and allowing for maintenance

- [ ] How to run tests? (looks pretty standard, `npm install && npm run test`)
- [ ] Expectations for contributions in `Contributing.md` - "test new functionality, run tests, etc."
- ... (will update as i continue to read)
