### 2024-02-19 neuroanatomy/thresholdmann#18 @sneakers-the-rat
[joss][tests] Test run failure - `before` hook times out
Part of: https://github.com/openjournals/joss-reviews/issues/6336

## Thank you for your feedback.

### Current behaviour

Trying to run tests:

```bash
npm install
npm run test
```

yields

```
> thresholdmann@1.0.0 test
> mocha --exit

  Test Thresholdmann
    1) "before all" hook in "Test Thresholdmann"

  0 passing (5s)
  1 failing

  1) Test Thresholdmann
       "before all" hook in "Test Thresholdmann":
     Error: Timeout of 5000ms exceeded. For async tests and hooks, ensure "done()" is called; if returning a Promise, ensure it resolves. (/Users/jonny/git/forks/thresholdmann/test.mjs)
      at listOnTimeout (node:internal/timers:573:17)
      at process.processTimers (node:internal/timers:514:7)

```

I suspect this has something to do with my use of lulu to block most traffic to google, as chrome attempts to launch and requests network access. I allowed it temporarily but that created a new issue:

```
  1) Test Thresholdmann
       "before all" hook in "Test Thresholdmann":
     Error: net::ERR_CONNECTION_REFUSED at http://127.0.0.1:8080
```

### Expected behaviour

Expected to run tests!

### Version information (for bug reports)

* **Browser + version**: Firefox 122.0.1
* **Your OS + version**: macOS 14.3.1
* npm: 10.2.4
* deps: (matching lockfile at https://github.com/neuroanatomy/thresholdmann/blob/25090ab07c7e51b4122f1ffb0997e8720d6b1daf/package-lock.json )
