### 2024-02-20 neuroanatomy/thresholdmann#20 @sneakers-the-rat
[joss][tests] Add test cases for listed functionality
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Ensure tests cover functionalities listed in `/doc.html`

- [x] Load and navigate data
- [x] Adjust threshold value - i'm not seeing adjusting the threshold per se, but I do see changing threshold settings like changing threshold direction. would be good to add that & compare to known output.
- [x] Select direction of thresholding
- [x] Move control point
- [x] Add control points
- [x] Select a control point
  - [x] Adjust threshold slider
- [x] Adjust mask opacity - not sure how you test the actual appearance of opacity, but should test get/set value
- [x] Adjust MRI contrast and brightness - same as ^
- [x] 3D render - the test for opening the page works for me :)
- [x] View threshold value
- [x] Saving
  - [x] Control points
  - [x] Mask
- [x] Loading Control points

I think most of this could be done by having a known-good mask for the demo input data at 2 or 3 configurations of points/thresholds. I think you may want to refactor into a `tests` directory to be able to group input/output data separately from the `img` directory, which seems to be the public `img` directory used on the site.

Authors lmk which of these are impossible/etc. i'm just recreating the list from `/doc.html` :)
