### 2024-02-20 neuroanatomy/thresholdmann#21 @sneakers-the-rat
[joss][package structure][nice to have] Refactor into `src`-like directoy structure?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Not a requirement of the review, but a nice-to-have:

Currently the project is structured s.t. the public assets/pages/etc. are in the same top-level directory as the tests and package metadata. Would be nice to make a clearer distinction between what is intended to be deployed on a web server/run locally vs. what is intended to be used as support. this is partially cosmetic, but would also be important for extensibility/maintainability later, as eg. adding more test cases that an http server would be serving could be undesirable if they exposed some vulnerability to the host machine.

So that would look something like

```
.eslintignore
.eslintrc.js
.gitignore
Contributing.md
LICENSE
package.json
...
src
  index.html
  thresholdmann.css
  thresholdmann.js
  ...
  img/
  render3D/
tests
  test.mjs
  fixtures.mjs
  data/
  ...
```
