### 2024-02-20 neuroanatomy/thresholdmann#22 @sneakers-the-rat
[joss][question] Size of $\epsilon$ term?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

The text reads

> where $\epsilon$ is a small number used to prevent a division by 0 at the exact position of control points.

I see the $\epsilon$ term is set as `0.001` here: https://github.com/neuroanatomy/thresholdmann/blob/25090ab07c7e51b4122f1ffb0997e8720d6b1daf/thresholdmann-worker.js#L29

Is that always a trivially small value, or is there some scale (that's reasonably likely in real data) where that might be a problem? I guess i am not quite clear on the units of the space to know. it seems like it's in voxel index space?

just posing as a question, y'all are the experts on brains of all sizes, and whether there might be some tiny hydra neuron edge case to worry about :)
