### 2024-02-20 neuroanatomy/thresholdmann#23 @sneakers-the-rat
[joss][packaging][question] CDN vs. webpacking?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

First, let me say that i always love seeing your work because I love how it is framework-free, vanilla JS that is just ready to run in the browser. So this question is certainly not me saying "you must do super heavy contemporary javascript stuff" but more thinking aloud about some of the tradeoffs.

So y'all have a chain of packages that are all good, but the dependency chain is also getting a little opaque and I am worrying a bit about maintainability - will y'all get to a point in 5 years where the connections between packages become too complex to handle?

For example:
- Depend on `mriviewerjs@v0.1.8` here: https://github.com/neuroanatomy/thresholdmann/blob/25090ab07c7e51b4122f1ffb0997e8720d6b1daf/index.html#L200
- Depend on `mrijs@0.0.6` here: https://github.com/r03ert0/mriviewerjs/blob/f937bed3f174417619c3426a0802096cc0d2dcc1/mriviewer.js#L7
- Depend on `structjs@0.0.1` here: https://github.com/neuroanatomy/mrijs/blob/dfa9fc8a195d137e67ba4c24f018717c19cebce9/mri.js#L10

Say an update is needed in one of those packages, how will y'all know all the downstream packages that need to get a patch release?

Another problem is that it becomes impossible to do what might be one of the best things about small-web tools like this, and that's use it offline/in low web resource contexts: one needs an internet connection and access to `cdn.jsdelivr.net` (and for that to be allowable by CORS).

I wonder how y'all feel about making these packages es6 (or whatever) packages so they could be added to project dependencies, and then use a very simple webpack configuration to create batteries-included, no network required distribution versions? Those would then be totally self-contained html files, would also allow you to reuse parts of the packages potentially more easily, and bump all dependencies up to the top level of the package for inspectability and maintenance.

I'd be happy to PR to show how simple that could be (and wouldn't require restructuring literally everything) if you want.

again this isn't required for the review, but me trying to think aloud about impacts of this structure (which works!) 5 or 10 years down the line :)
