### 2024-02-20 neuroanatomy/thresholdmann#24 @sneakers-the-rat
[joss] Handling non-8bit images?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

I can't tell if this is a bug or not.

It looks like the expectation for 8-bit images is encoded in a few places (ctrl+f for `255`), and I'm wondering if that's a problem for images that aren't 8-bit.

I created a version of the demo data cast to `uint16`:

```python
from nibabel import load
from nibabel.nifti1 import save, Nifti1Image
import numpy as np

img = load('test-image.nii.gz')
data = img.get_fdata()
data16 = data.astype(np.uint16) * 256

img16 = Nifti1Image(data16, np.eye(4))
save(img16, 'test-image-16.nii.gz')
```

and it looks like the image is correctly loaded, and I'm able to do all the thresholding correctly as expected.

The threshold mask shouldn't be a problem since that's just a boolean array, but I do notice that `control-points.json` takes values from 0-255. So it seems like the image is downscaled to 8-bit for use internally within thresholdmann, but that might pose a problem for someone who uses a different bit depth, since they'll have to convert the ranges to the source range.

How hard would it be to keep the values in the original range of the image, either internally in the program, or recast at export time? or is this not a problem?
