### 2024-02-20 neuroanatomy/thresholdmann#26 @sneakers-the-rat
[joss][ui] Restrict control points to be within the borders of the image
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Ok this one i find extremely charming.

You can drag the points outside the image, onto the menus, and it... works.

https://github.com/neuroanatomy/thresholdmann/assets/12961499/3d89eb7a-491b-4f06-b6d2-488568827519

So as much as i literally love that (bravo on making the code robust enough to just casually handle that, the points don't even disappear behind the menus), it would probably not be great to be able to do that since it would create some undefined behavior in downstream uses of the control points (what does it mean for the point to be in negative voxel space, or in a voxel larger than the size of the dimension?)
