### 2024-03-18 #6336 @katjaq
Hi @adamltyson . Thank you so much for the ping. <3
At first, we were waiting a bit for the reviews of @anibalsolon @62442katieb – but then we just went ahead and started tackling the issues raised in the awesome review by @sneakers-the-rat 🌟 and implemented the suggested improvements. We will keep pushing our updates in the coming days :)
Thanks again for all your great comments @sneakers-the-rat . <3
