### 2024-03-19 neuroanatomy/thresholdmann#20 @katjaq
Thank you for your suggestions @sneakers-the-rat ! All the tests have now been implemented and merged in the PR mentioned above #29 . Thank you @ntraut !

We will let you check it out and look forward to your reply or approval :)
