### 2024-03-19 neuroanatomy/thresholdmann#16 @katjaq
Thank you for the suggestion!
We implemented it in commit [58f8f09](https://github.com/neuroanatomy/thresholdmann/commit/58f8f092f28fb7e9bd4b0715ed4908ed4b5bc03d)
Just one demo dataset – but we hope it's ok :)
