### 2024-03-19 neuroanatomy/thresholdmann#22 @katjaq
Thank you for pointing this out.

It is indeed in voxel space, and with current MRI data, we would not expect this to become a problem. It does change the exact space at which a control point faints out, but in practice, the amount of voxels in an MRI even if it were smaller or bigger, is usually about 300. We also tested with high resolution data of small brains (for example a mouse: dim: 420 x 548 x 282, pixdim: 0.03 x 0.03 x 0.03) and also with a big dolphin brain (dim: 461 x 424 x 304, pixdim: 0.30 x 0.30 x 0.30) and the behaviour of the tool doesn't perceivably change.

But we could change it in future if we realise that it becomes a problem, so that a user could select a value.
