### 2024-03-19 neuroanatomy/thresholdmann#24 @katjaq
That's a very good question. We don't actually do anything with the user's raw MRI data – they just provide it to the app so they can create a mask for their dataset. The MRI data is handled by mriviewerjs which handles the different formats – so your modified demo data was loaded without problem. For the mask, it should be all fine. For the control-points, their values are displayed as integers in the interface, but upon saving are saved as floats. So we think it's not a problem.
