### 2024-03-19 neuroanatomy/thresholdmann#26 @katjaq
:D Thanks for spotting this! And for your video for illustration (and eternity :D)
May we call it a feature ? 🙃  It is like any control point – just "far away by a certain amount" and hence the values will be handled accordingly. :) Voxels outside the brain, just negative coordinate, don't mean anything.
:D The points don't disappear behind the menu and you can even catch the point once it is outside, you can bring it back in ...
... ya, you are right, we should probably restrict the space. 🙈
