### 2024-03-19 neuroanatomy/thresholdmann#26 @sneakers-the-rat
I'll leave this one up to you - it is one of those things that's like... probably fine? and you're right, it is technically valid as far as setting a parametric threshold map goes, since the 0 point is arbitrary anyway? and it doesn't break the interface, and the tool handles it just fine, so i think the only concern is downstream usage? so maybe a tickbox to "constrain points within volume" or something for if that becomes a problem for ppl lol
