### 2024-03-20 neuroanatomy/thresholdmann#26 @katjaq
Thanks @sneakers-the-rat . <3 We have now restricted control points to remain within the bounding box in PR #32 – they will stay inside, closest to the cursor, in case it leaves the data.
