### 2024-03-21 neuroanatomy/thresholdmann#21 @katjaq
Thank you for pointing this out and your suggestion. We were actually happy to do this at first, but then we realised it is maybe better in our case to keep it this way, and should not be too problematic?
We're currently serving thresholdmann from `/(root)`. Checking the Github Pages settings, there doesn't seem to be an option to serve the website from a folder that wouldn't be either `/(root)` or `/docs`. We could keep only `index.html` at `/(root)`, all the website files at `/src`, and all the test files at `/tests`. That would work around this issue, but defeat the purpose of the restructuring. 🙈 We also would like to avoid links to our tool such as `https://neuroanatomy.github.io/thresholdmann/src/`. So, weighting the benefits and possible implications, being happy with serving from Github pages, we would prefer to keep the current structure of the code. Does it sound reasonable @sneakers-the-rat ?
