### 2024-03-21 neuroanatomy/thresholdmann#23 @katjaq
If you agree with our implementations based on your other issues and the current code structure, we think this should now be the (for the moment) final code and we would be very happy if your proposition still stands to show us how this could be packaged  and work this way <3
