+++ config/webpack.common.js
@@ -1,0 +1,80 @@
+const CopyWebpackPlugin = require('copy-webpack-plugin')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
+const webpack = require("webpack");
+const ImportHttpWebpackPlugin = require('import-http/webpack')
+
+const paths = require('./paths')
+
+module.exports = {
+  // Where webpack looks to start building the bundle
+  entry: {
+    index: paths.src + '/index.js',
+    // render3D: paths.src + '/render3d/index.js'
+  },
+
+  // Where webpack outputs the assets and bundles
+  output: {
+    path: paths.build,
+    filename: '[name].bundle.js',
+    publicPath: '/',
+    clean: true,
+    // module: true
+  },
+
+  // Customize the webpack build process
+  plugins: [
+    // Copies files from target to destination folder
+    new CopyWebpackPlugin({
+      patterns: [
+        {
+          from: paths.img,
+          to: 'img',
+          globOptions: {
+            ignore: ['*.DS_Store'],
+          },
+          noErrorOnMissing: true,
+        },
+      ],
+    }),
+
+    // Generates an HTML file from a template
+    // Generates deprecation warning: https://github.com/jantimon/html-webpack-plugin/issues/1501
+    new HtmlWebpackPlugin({
+      title: 'Thresholdmann',
+      template: paths.src + '/index.html', // template file
+      filename: 'index.html', // output file
+      scriptLoading: "blocking",
+      inject: "head"
+    }),
+
+    new webpack.ProvidePlugin({
+      $: 'jquery',
+      jQuery: 'jquery',
+    }),
+
+    new ImportHttpWebpackPlugin()
+  ],
+
+  // Determine how modules within the project are treated
+  module: {
+    rules: [
+      // JavaScript: Use Babel to transpile JavaScript files
+      { test: /^\.js$/, use: ['babel-loader'] },
+
+      // Images: Copy image files to build folder
+      { test: /\.(?:ico|gif|png|jpg|jpeg)$/i, type: 'asset/resource' },
+
+      // Fonts and SVGs: Inline files
+      { test: /\.(woff(2)?|eot|ttf|otf|svg|)$/, type: 'asset/inline' },
+    ],
+  },
+
+  resolve: {
+    modules: [paths.src, 'node_modules'],
+    extensions: ['.js', '.jsx', '.json'],
+  },
+
+  // experiments: {
+  //   outputModule: true,
+  // },
+}
