+++ config/webpack.prod.js
@@ -7,7 +7,7 @@ const common = require('./webpack.common')
 
 module.exports = merge(common, {
   mode: 'production',
-  devtool: false,
+  devtool: 'source-map',
   output: {
     path: paths.build,
     publicPath: '/',
