### 2024-04-03 neuroanatomy/thresholdmann#17 @sneakers-the-rat
Fabulous! That's pretty much all i came up with as i was reading! feel free to close
