### 2024-04-03 neuroanatomy/thresholdmann#20 @sneakers-the-rat
best test description of all time. shouldn't all code? <3  https://github.com/neuroanatomy/thresholdmann/blob/9861286089b7b0ec9b89b77a5b50971820613440/test.mjs#L38
