### 2024-04-03 neuroanatomy/thresholdmann#20 @sneakers-the-rat
Looks great! thanks for working on this. hopefully that helps with contribs and gives you one less thing to lose sleep over in the future of this tool :)
