### 2024-04-03 neuroanatomy/thresholdmann#21 @sneakers-the-rat
ooh yes! this is another thing that the webpack stuff can help with - since gh pages are just gh actions now, you can make an action that builds the site to a branch and then deploys from there so you have more control over the structure of your source directory.

> We also would like to avoid links to our tool such as https://neuroanatomy.github.io/thresholdmann/src/.

most definitely a higher priority!!!!

this was mostly just a "would this be nice for y'all?" issue, so no worries about keeping it this way.

here's a super trivial example of building something with sphinx and deploying to gh pages - https://github.com/p2p-ld/docs/blob/main/.github/workflows/sphinx.yml

yours would be even simpler since it's ready-to-roll JS

```yaml
name: Deploy
on:
  push:
    branches:
      - main

jobs:
  deploy:
    runs-on: ubuntu-latest
    permissions:
      contents: write  # To push a branch
    steps:
      - uses: actions/checkout@v3
      - name: deploy to gh-pages
        uses: JamesIves/github-pages-deploy-action@v4
        with:
          folder: src
```

but again only do that if it's easier for you to tend to your code!
