### 2024-04-03 neuroanatomy/thresholdmann#34 @sneakers-the-rat
[joss][docs] Performance limits?
Part of: https://github.com/openjournals/joss-reviews/issues/6336

one more thing - I assume there are some images that would be so big they would crash the browser/etc. have y'all done any sort of performance benchmarking to give people interested in this tool some idea of "this tool is for you if your images are <1TB big"

It's a documentation question rather than a capability one - it works great with the sample images i threw at it, and i wouldn't expect to use it for those humongous bazillion Tesla scans that need a whole room of hard drives to store, so mostly a "would be nice" section in the docs to pre-empt the inevitable issue of someone asking if it's for them!
