### 2024-04-03 #6336 @sneakers-the-rat
Finished with my review! My checklist updated with my final comments, pasting here to they're also in the thread of conversation. Thanks again to the authors for their work and for addressing all my questions and comments.

> If there is one thing you can always say about @katjaq , @r03ert0 , and @ntraut 's code it's that it does exactly what it says it does with style, simplicity, and solidity. Thanks to the authors for this lovely tool. I am only familiar with working with fMRI data being a huge hassle of gigantic packages and cumbersome GUIs, and so this tool that you can hold in the palm of your hand and run anywhere is an absolute treat. I am a huge fan of the author's small web, single purpose tool philosophy, and this is another great entry in that saga.
>
> I have said this a number of times in the issues, but I think this kind of web development is especially lovely to see in an era of increasingly bloated javascript frameworks that are difficult to maintain and reverse engineer. for this tool to literally be deployed as is from a static web server - not even a build stage! - shows how much more legroom the web browser as programming environment has. The code is capable, concise, and correct. very well done. Docs are good, tests are good, all claims made in the paper easily validated, fundamentals are sound.
>
> The area of growth i have suggested for the authors is to use *some* of those tools to extend the functional modularity of their work into a bit more fluid programmatic modularity. They make excellent use of their prior work with mriviewer and muijs, but the use of global page state does put a little bit of a cap on how the excellent work they have done here can be reused. One of the virtues of their very modular approach is that it naturally lends itself to compositionality - it's already possible to mix and match their tools as part of a researcher's workflow, the next level would be to be able to mix and match them in downstream tools. I have written a skeleton draft of a transitional path there in an above linked PR for the authors to take or leave. Even without tools like webpack, ES6 modules are [well supported](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules#browser_compatibility) by browsers, so the authors can keep the dream of the low-fi but high-function read/write web alive *and* make a lovely little ecosystem that builds and builds and builds on itself!
>
> I don't have much more to say since the work speaks for itself. Thanks again to the authors for their work, i will continue to jump at the chance to use and look at any work y'all do - always a treat. They have addressed all my suggestions and i wholeheartedly endorse this for JOSS!
