### 2024-04-09 a061d22
add information on epsilon term
