### 2024-04-09 neuroanatomy/thresholdmann#22 @katjaq
documented! In commit [a061d22](https://github.com/neuroanatomy/thresholdmann/commit/a061d222c34d1954ec79dabb9406d3c29573c53e) Thank you again :)
