### 2024-04-09 neuroanatomy/thresholdmann#23 @katjaq
Thanks a lot for this draft PR @sneakers-the-rat . We are learning so much with you! 🥰 We will look into it. Thank you for not requiring this for the JOSS review and giving us the opportunity to continue this effort independent from it. <3
