### 2024-04-09 neuroanatomy/thresholdmann#21 @katjaq
Thank you very much @sneakers-the-rat . This sounds good. For the moment, we would like to keep things as they are, and once we will look into the PR that you so kindly drafted for starting the packaging, moving to actions here will become interesting.
