### 2024-04-09 neuroanatomy/thresholdmann#23 @sneakers-the-rat
Of course!!! The only things I would "require" are the things on the checklist, the rest are just ideas and suggestions up to your discretion - its your work, you know it better than me :)
