### 2024-04-09 #6336 @katjaq
Yeyyy! We completed answering all changes as suggested by @sneakers-the-rat ! That was an amazing review experience.

We thank @sneakers-the-rat for all the great comments, suggestions and code contributions which have made our tool more robust, well documented, user- and contributor friendly and apt for being integrated into neuroscientific workflows. We answered the issues raised, implemented tests to cover all functionalities, include a demo dataset for quick tool exploration, improved user- and code documentation, have a test setup that works across many browsers, restricted control points to the  data boundaries, and tested performance on an unusually large dataset. We thank the reviewer for showing us how to move towards programmatic modularity in the future with web packaging and refactoring our code to separate source files from the rest and deploy using GitHub actions.
