### 2024-04-19 neuroanatomy/thresholdmann#36 @anibalsolon
[joss][ui] controls become unusable upon text selection
Part of: https://github.com/openjournals/joss-reviews/issues/6336

when mistakenly selecting the text (and controllers), the sliders stop working.

![image](https://github.com/neuroanatomy/thresholdmann/assets/562525/ed09586e-ccc9-4fe4-ba73-35952792e84d)

I'd suggest to apply the following CSS rules to disable text selection on the controllers and re-enabling it to the info panel.
```css
#buttons {
    user-select: none;
}
#info {
    user-select: text;
}
#control {
    user-select: none;
}
```
