### 2024-04-19 neuroanatomy/thresholdmann#37 @anibalsolon
[joss][ui] easier point moving
Part of: https://github.com/openjournals/joss-reviews/issues/6336

A suggestion: Unless it is a common user case to add points very closely, I believe dragging and dropping could always be enabled, thus removing the 'Move' mode.

![image](https://github.com/neuroanatomy/thresholdmann/assets/562525/097aaf20-f6c2-4f49-b22c-60225821ef5f)
