### 2024-04-19 neuroanatomy/thresholdmann#38 @anibalsolon
[joss][ui] numeric input
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Currently, the threshold inputs are of type `text`. Given you are providing `min` and `max`, I assume it will always be a number. I'd suggest to change the field type to "number", so 1) the range is applied, and 2) no other values are provided (e.g. characters).

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/number

![image](https://github.com/neuroanatomy/thresholdmann/assets/562525/63cd4a34-dae9-4f4b-ab6b-5143a726495e)
