### 2024-04-19 neuroanatomy/thresholdmann#39 @anibalsolon
[joss][sec] possible injection and data leakage
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Loading a malicious JSON file can execute code and leak MRI data (or worse). A simple validation for numeric values during loading can prevent such scenario.

```json
{"points":[[100,164,"0\"><script>window.alert('hey jude')</script><br style=\"display: none;\" data-close=\""]],"values":[10]}
```
