### 2024-04-19 neuroanatomy/thresholdmann#40 @anibalsolon
[joss][feature] negative selection & flood fill
Part of: https://github.com/openjournals/joss-reviews/issues/6336

Have you folks crossed through the need of control points that remove such range of values?

In the same topic, have you considered using flood fill ("magic wand") to perform selection of some brain structures, while maintaining others intact?
