### 2024-04-23 #6336 @adamltyson
Hi @katjaq, this looks ready to go :tada:

Could you create a new tagged release of the software, archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or other), and post the version number and archive DOI here? I'll then get the ball rolling with the rest.
