### 2024-04-27 4f4185b
improve control point visibility
