### 2024-04-27 neuroanatomy/thresholdmann#35 @katjaq
Thank you for this feedback and the linked resource which will keep in mind also for future applications :)
We implemented a thin solid white border around each control point for better visibility in commit #[4f4185b](https://github.com/neuroanatomy/thresholdmann/commit/4f4185bb7e31d55c63be2d355e09a311594314ff) as suggested by the reviewer.
