### 2024-04-27 344f987
disable text selection for ui controls
