### 2024-04-27 neuroanatomy/thresholdmann#36 @katjaq
Thank you for the feedback and the improvement fix @anibalsolon . ⭐
We implemented the solution suggested in commit #[344f987](https://github.com/neuroanatomy/thresholdmann/commit/344f98746781f6f8778e05d5d2e405e69cca3086).
