### 2024-04-27 ffbda69
set threshold value to numeric input
