### 2024-04-27 neuroanatomy/thresholdmann#38 @katjaq
Thank you for the suggestion. We implemented this change of input type from text to number in commit #[ffbda69](https://github.com/neuroanatomy/thresholdmann/commit/ffbda696ab091556168cc0d2dcc659393436a1a6)
