### 2024-04-27 neuroanatomy/thresholdmann#37 @katjaq
Thank you for the suggestion. We thought about it , but we prefer to keep these 4 options separated. Each of them requires just one key press to be active, out of only 4 options, which shouldn't be in the way of a fluid working process.
The 4 options allow for a quick select of the control point to then change its value; if moving were enabled this could potentially lead to unexpected behaviour. It is also a nice use case to place control points close to each other to obtain boundaries – for example one to exclude the meninges; one to include the cortex.
