### 2024-04-27 neuroanatomy/thresholdmann#40 @katjaq
We hadn't thought about flood fill. It may be nice to have in some cases. We have thought, however, about adding some basic mathematical morphology tools (erosion, dilation, closing, opening, connected components, etc.). At the moment, we don't have an efficient implementation, and we do all that type of post-processing using scikit-image, in Python.
