### 2024-04-27 neuroanatomy/thresholdmann#39 @katjaq
Thank you for the note. We hadn't thought about this as usually the user themselves is the producer of the json file they'll later load. We'll integrate a sanitizer.
