### 2024-04-27 neuroanatomy/thresholdmann#39 @katjaq
A check for numbers only has been implemented in # [6bdaf8c](https://github.com/neuroanatomy/thresholdmann/commit/6bdaf8cb84d9b4f00423263ff56400630a5b361f). Thank you @r03ert0 .
