### 2024-04-27 78f73ff
fix NaN finding procedure
