+++ ReadMe.md
@@ -5,6 +5,7 @@ A Web tool for interactively creating adaptive thresholds to segment MRI data.
 Katja Heuer, Nicolas Traut & Roberto Toro, November 2023
 
 [![CircleCI](https://circleci.com/gh/neuroanatomy/thresholdmann.svg?style=shield)](https://circleci.com/gh/neuroanatomy/thresholdmann)
+[![DOI](https://zenodo.org/badge/135712332.svg)](https://zenodo.org/doi/10.5281/zenodo.11080336)
 
 
 Simply drag & drop your MRI file for display in an interactive stereotaxic viewer. Move and add a control point and grow or shrink your selection from there.
