### 2024-04-28 #6336 @katjaq
Dear @adamltyson . We merged all our changes from the review process from master into the JOSS paper branch. As no changes to the paper were suggested, the compiled version you provided above is unchanged and we confirm its layout and the generated pdf.

We checked all the items in the list you gave us. Here is the information:

- [x] Double check authors and affiliations (including ORCIDs) – all correct
- [x] Version number of the joss paper release: [v.1.0.0](https://github.com/neuroanatomy/thresholdmann/releases/tag/v.1.0.0)
- [x] DOI from the corresponding Zenodo archive: [10.5281/zenodo.11080336](https://zenodo.org/doi/10.5281/zenodo.11080336)
