### 2024-04-28 439d7bf
Update the way tests are implemented now after the review process
Update paper.md
