+++ paper.md
@@ -34,7 +34,7 @@ Brain extraction and segmentation are required for most analyses of neuroimaging
 Thresholdmann complements the variety of existing brain segmentation tools, providing an easy interface to manually control the segmentation on a local scale across different brain imaging modalities and image contrast gradients. The masks produced by Thresholdmann can serve as a starting point for more detailed manual editing using tools such as BrainBox ([https://brainbox.pasteur.fr](https://brainbox.pasteur.fr)) [@BrainBox] or ITK-SNAP ([http://www.itksnap.org](http://www.itksnap.org)) [@itkSNAP]. This interactive approach is especially valuable for non-human brain imaging data, where automatic approaches often require extensive manual adjustment anyway [@PrimeDE; @PrimeRE; @PrimeDE2]. We have used Thresholdmann successfully to create initial brain masks for a variety of vertebrate brains – including many non-human primate datasets [@34primates; @cerebella] – as well as developmental data. Small Web tools, such as Thresholdmann or Reorient ([https://neuroanatomy.github.io/reorient](https://neuroanatomy.github.io/reorient)) [@Reorient], focused on solving a single problem, can become helpful additions to the methodological toolbox of neuroimagers.
 
 # Methods
-The spatially-varying threshold is computed from a number of control points. Each control point has a position $x_i$ and a threshold value $v_i$ that can be adjusted interactively. At each point $x$ of the volume, the local threshold is computed as a weighted function of the control points. The weight associated to the i-th control point at position $x$ is given by:  
+The spatially-varying threshold is computed from a number of control points. Each control point has a position $x_i$ and a threshold value $v_i$ that can be adjusted interactively. At each point $x$ of the volume, the local threshold is computed as a weighted function of the control points. The weight associated to the $i$-th control point at position $x$ is given by:  
 $$w_i \left( x \right) = \dfrac1 {\left( x-x_i \right)^2+\epsilon} ,$$  
 where $\epsilon$ is a small number used to prevent a division by 0 at the exact position of control points. The value of the threshold at position $x$ is the obtained as:  
 $$v \left( x \right) = \dfrac{ \sum\limits_i^n v_i w_i \left( x \right)} {W \left( x \right)} ,$$  
@@ -43,7 +43,7 @@ $$W \left( x \right) = \sum_i^n w_i \left( x \right) .$$
 
 Additionally, we add a “background” threshold value which makes the action of the control points localised. The background value is independent of the distance and is given a small constant weight. In that manner, positions far from all control points get the background threshold value.  
 
-Thresholdmann was coded in JavaScript and runs from a GitHub page. Code style was verified using ESLint ([https://eslint.org](https://eslint.org)). Unit tests and end-to-end tests were implemented using Mocha ([https://mochajs.org](https://mochajs.org)) and Puppeteer ([https://pptr.dev](https://pptr.dev)). Modifications in the code are continuously tested using CircleCI ([https://circleci.com](https://circleci.com)).
+Thresholdmann was coded in JavaScript and runs from a GitHub page. Code style was verified using ESLint ([https://eslint.org](https://eslint.org)). Unit tests and end-to-end tests were implemented using Mocha ([https://mochajs.org](https://mochajs.org)) and Playwright ([https://playwright.dev](https://playwright.dev)). Modifications in the code are continuously tested using CircleCI ([https://circleci.com](https://circleci.com)).
 
 # Figures
[…]
