### 2024-04-29 #6336 @adamltyson
Hi @katjaq, based on [this](https://github.com/openjournals/joss-reviews/issues/5358#issuecomment-1663637277) and [this](https://github.com/openjournals/joss-reviews/issues/5492#issuecomment-1774830341) comment, it seems as if setting a fraction in displaystyle using `\dfrac` is not supported by the JOSS compilation tool. Could you change it to `\frac` and see if that fixes the problem?
