### 2024-04-29 f10b3cc
replace dfrac by frac
Update paper.md
