### 2024-04-29 #6336 @katjaq
Hello @adamltyson . Yeyyyyy, thank you so much for setting publication in motion 😍
I changed it to `\dfrac` and it seems to look like before from what I can see in the GitHub actions :). So, 🤞 .  Sorry for the hiccup.
