### 2024-05-01 #6336 @Kevin-Mattheus-Moerman
@katjaq as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems to be in order, the only thing I've noticed is that your archive page does not list the license. Can you check (but attempting to manually edit the listing) if you can add your software license? Here is one for a different project, and as you can see it will be listed under "Rights": https://zenodo.org/records/10278296.
