### 2024-05-02 #6336 @katjaq
Hello @Kevin-Mattheus-Moerman . Thank you very much for your feedback.

The license is now listed in our [Zenodo entry](https://zenodo.org/records/11080337) :)
